api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.8.0rc2
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.13.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
oauthlib==3.2.2
openai==1.16.2
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tqdm==4.66.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
zope.interface==6.2
